requirements.txt
pytest==6.0.0
requests==2.25.1
jsonschema==3.2.0
python-dotenv==0.17.1